fix: move marketplace to server wrapper to fix force-dynamic in CI build
Route segment config (export const dynamic) is only recognized in Server
Components. The previous placement inside a 'use client' file was silently
ignored, so Next.js still attempted static prerendering and hit
ECONNREFUSED in CI.

Split into:
- page.tsx: Server Component with force-dynamic + re-export
- MarketplaceClient.tsx: 'use client' component with all existing logic

No behavior changes; all 67 tests pass.
